Prevent Escape from canceling agent while overlays are open
