+++ apps/twig/src/renderer/features/sessions/components/SessionView.tsx
@@ -11,7 +11,6 @@ import {
   usePendingPermissionsForTask,
 } from "@features/sessions/stores/sessionStore";
 import type { Plan } from "@features/sessions/types";
-import { useSettingsDialogStore } from "@features/settings/stores/settingsDialogStore";
 import { useSettingsStore } from "@features/settings/stores/settingsStore";
 import { useAutoFocusOnTyping } from "@hooks/useAutoFocusOnTyping";
 import { Spinner, Warning } from "@phosphor-icons/react";
@@ -21,8 +20,6 @@ import {
   isJsonRpcNotification,
   isJsonRpcResponse,
 } from "@shared/types/session-events";
-import { useCommandMenuStore } from "@stores/commandMenuStore";
-import { useShortcutsSheetStore } from "@stores/shortcutsSheetStore";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { useHotkeys } from "react-hotkeys-hook";
 import { getSessionService } from "../service/service";
@@ -139,19 +136,6 @@ export function SessionView({
     disabled: !isRunning,
     isLoading: !!isPromptPending,
   });
-
-  const isSettingsOpen = useSettingsDialogStore((s) => s.isOpen);
-  const isCommandMenuOpen = useCommandMenuStore((s) => s.isOpen);
-  const isShortcutsSheetOpen = useShortcutsSheetStore((s) => s.isOpen);
-  const hasOverlay =
-    isSettingsOpen || isCommandMenuOpen || isShortcutsSheetOpen;
-
-  useHotkeys(
-    "escape",
-    onCancelPrompt,
-    { enabled: !!isPromptPending && !hasOverlay },
-    [onCancelPrompt],
-  );
 
   useHotkeys(
     "shift+tab",
